Fix crash with a workaround, improve performance
ase.kpoints was crashing if the "height" of the hexagonal cell was smaller
then the "side", now the cell is temporarly tall, after the kpoints
bandpath calculation the size returns to a normal value.
The height of the cell during the bandpath sampling doesn't matter
because the path is on the x-y plane.

Improve performance using ScalaPack parallelization and band/kpoints
parallelization, this also seems to solve a MemoryError problem, at
least for now.

Other minor modifications to printed values and file names.