Fix band structure, remove useless scripts
ASE can write to JSON and plot bandstructure out of the box, no need to
write new scripts.
Finally fix every problem with the band structure, the band path is
NOT computed using lattice functions but with the bandpath() external
function.
Band structure is obtained with the external function
get_band_structure() in order to skip the lattice check from ASE.
Every problem with the previous code was caused by a bad behaviour of
niggli_reduce() inside the ASE code, when the hexagonal cell is too tall
the vectors are mixed and then every check on the cell will fail.
Save everything in a subfolder, not to create a mess.